@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.3 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,29 +29,34 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  collection: ^1.16.0
-  intl: ^0.17.0
-  dynamic_color: ^1.4.0
+  collection: ^1.17.1
+  intl: ^0.18.1
+  dynamic_color: ^1.6.5
   cupertino_icons: ^1.0.5
-  latlong2: ^0.8.1
-  geolocator: ^9.0.1
-  flutter_riverpod: ^2.0.0-dev.9
+  latlong2: ^0.9.0
+  geolocator: ^9.0.2
+  flutter_riverpod: ^2.3.6
+  # consider upgrading to sliding_up_panel2
   sliding_up_panel: ^2.0.0+1
-  flutter_platform_widgets: ^2.0.0
-  dio: ^4.0.6
-  path_provider: ^2.0.11
-  fl_chart: ^0.55.0
-  flutter_dotenv: ^5.0.2
-  image_picker: ^0.8.5+3
-  flutter_image_compress: ^1.1.1
-  flutter_map: ^2.2.0
-  flutter_map_tile_caching: ^5.0.0
-  flutter_map_location_marker: ^4.1.3
-  flutter_map_marker_popup: ^3.2.0
-  # updates for flutter map 2
-  flutter_map_tappable_polyline:
-    git: https://github.com/trilliumlab/flutter_map_tappable_polyline.git
+  flutter_platform_widgets: ^3.3.5
+  dio: ^5.2.1+1
+  path_provider: ^2.0.15
+  fl_chart: ^0.63.0
+  flutter_dotenv: ^5.1.0
+  image_picker: ^0.8.9
+  flutter_image_compress: ^2.0.3
+  flutter_map: ^5.0.0
+  flutter_map_tile_caching: ^9.0.0-dev.2
+  flutter_map_location_marker: ^7.0.0
+  # TODO need to get this working
+  # flutter_map_marker_popup: ^5.1.0
+  flutter_map_tappable_polyline: ^5.0.0
+#  # updates for flutter map 2
+#  flutter_map_tappable_polyline:
+#    git: https://github.com/trilliumlab/flutter_map_tappable_polyline.git
+
   # fixes elevation not being read
+  # TODO update to latest version
   gpx:
     git: https://github.com/trilliumlab/dart-gpx.git
   # allow simplification of LatLng
