@@ -11,7 +11,7 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await dotenv.load();
-  FlutterMapTileCaching.initialise(await RootDirectory.temporaryCache);
+  await FMTC.initialise();
   await FMTC.instance('forestPark').manage.createAsync();
   runApp(const App());
 }
