@@ -1,11 +1,10 @@
 import 'package:dio/dio.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:forest_park_reports/consts.dart';
 
 final dioProvider = Provider((ref) {
-  const baseUrl = "https://forestpark.elliotnash.org/api/v1/";
-  // const baseUrl = "http://192.168.0.100:8000/api/v1";
   final options = BaseOptions(
-    baseUrl: baseUrl
+    baseUrl: kApiUrl
   );
   return Dio(options);
 });
